Move VerticalAlignmentConverter into its own file under the Converters namespace
