+++ MaterialDesignThemes.Wpf/SmartHint.cs
@@ -1,6 +1,4 @@
 ﻿using System.ComponentModel;
-using System.Globalization;
-using System.Windows.Data;
 using MaterialDesignThemes.Wpf.Converters;
 
 namespace MaterialDesignThemes.Wpf;
@@ -236,14 +234,4 @@ public class SmartHint : Control
             Dispatcher.BeginInvoke(action);
         }
     }
-}
-
-public class VerticalAlignmentConverter : IValueConverter
-{
-    public VerticalAlignment StretchReplacement { get; set; } = VerticalAlignment.Top;
-    public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-        => value is VerticalAlignment.Stretch ? StretchReplacement : value;
-
-    public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-        => throw new NotImplementedException();
 }
